requirements.txt
pytest==8.3.5
requests==2.32.3
websocket-client==1.8.0
python-dotenv==1.0.1
allure-pytest==2.13.5